refactor(web): type dsl import hotkey
