+++ web/app/components/app/create-from-dsl-modal/types.ts
@@ -1,0 +1,7 @@
+export const CreateFromDSLModalTab = {
+  FROM_FILE: 'from-file',
+  FROM_URL: 'from-url',
+} as const
+
+export type CreateFromDSLModalTab =
+  (typeof CreateFromDSLModalTab)[keyof typeof CreateFromDSLModalTab]
